relay_substrate_client

Enum Error

pub enum Error {

    Io(Error),
    RpcError(ClientError),
    ResponseParseFailed(Error),
    ChannelError(String),
    MissingRequiredParachainHead(ParaId, u64),
    FinalityProofNotFound(u64),
    ClientNotSynced(Health),
    FailedToGetSystemHealth {
        chain: String,
        error: Box<Error>,
    },
    FailedToReadBestFinalizedHeaderHash {
        chain: String,
        error: Box<Error>,
    },
    FailedToReadBestHeader {
        chain: String,
        error: Box<Error>,
    },
    FailedToReadHeaderHashByNumber {
        chain: String,
        number: String,
        error: Box<Error>,
    },
    FailedToReadHeaderByHash {
        chain: String,
        hash: String,
        error: Box<Error>,
    },
    FailedToReadBlockByHash {
        chain: String,
        hash: String,
        error: Box<Error>,
    },
    FailedToReadStorageValue {
        chain: String,
        hash: String,
        key: StorageKey,
        error: Box<Error>,
    },
    FailedToReadRuntimeVersion {
        chain: String,
        error: Box<Error>,
    },
    FailedToGetPendingExtrinsics {
        chain: String,
        error: Box<Error>,
    },
    FailedToSubmitTransaction {
        chain: String,
        error: Box<Error>,
    },
    FailedStateCall {
        chain: String,
        hash: String,
        method: String,
        arguments: Bytes,
        error: Box<Error>,
    },
    FailedToProveStorage {
        chain: String,
        hash: String,
        storage_keys: Vec<StorageKey>,
        error: Box<Error>,
    },
    FailedToSubscribeBestHeaders {
        chain: String,
        error: Box<Error>,
    },
    FailedToSubscribeFinalizedHeaders {
        chain: String,
        error: Box<Error>,
    },
    FailedToSubscribeJustifications {
        chain: String,
        error: Box<Error>,
    },
    UnorderedFinalizedHeaders {
        chain: String,
        prev_number: String,
        next_number: String,
    },
    BridgePalletIsHalted,
    BridgePalletIsNotInitialized,
    TransactionInvalid(TransactionValidityError),
    WaitingForRuntimeUpgrade {
        chain: String,
        expected: SimpleRuntimeVersion,
        actual: SimpleRuntimeVersion,
    },
    FinalityProofWeightLimitExceeded {
        extras: SubmitFinalityProofCallExtras,
    },
    Custom(String),

}

Errors that can occur only when interacting with a Substrate node through RPC.

Variants

Io(Error)

IO error.

RpcError(ClientError)

An error that can occur when making a request to an JSON-RPC server.

ResponseParseFailed(Error)

The response from the server could not be SCALE decoded.

ChannelError(String)

Internal channel error - communication channel is either closed, or full. It can be solved with reconnect.

MissingRequiredParachainHead(ParaId, u64)

Required parachain head is not present at the relay chain.

FinalityProofNotFound(u64)

Failed to find finality proof for the given header.

ClientNotSynced(Health)

The client we’re connected to is not synced, so we can’t rely on its state.

FailedToGetSystemHealth

Failed to get system health.

Fields

chain: String

Name of the chain where the error has happened.

error: Box<Error>

Underlying error.

FailedToReadBestFinalizedHeaderHash

Failed to read best finalized header hash from given chain.

Fields

chain: String

Name of the chain where the error has happened.

error: Box<Error>

Underlying error.

FailedToReadBestHeader

Failed to read best finalized header from given chain.

Fields

chain: String

Name of the chain where the error has happened.

error: Box<Error>

Underlying error.

FailedToReadHeaderHashByNumber

Failed to read header hash by number from given chain.

Fields

chain: String

Name of the chain where the error has happened.

number: String

Number of the header we’ve tried to read.

error: Box<Error>

Underlying error.

FailedToReadHeaderByHash

Failed to read header by hash from given chain.

Fields

chain: String

Name of the chain where the error has happened.

hash: String

Hash of the header we’ve tried to read.

error: Box<Error>

Underlying error.

FailedToReadBlockByHash

Failed to read block by hash from given chain.

Fields

chain: String

Name of the chain where the error has happened.

hash: String

Hash of the header we’ve tried to read.

error: Box<Error>

Underlying error.

FailedToReadStorageValue

Failed to read sotrage value at given chain.

Fields

chain: String

Name of the chain where the error has happened.

hash: String

Hash of the block we’ve tried to read value from.

key: StorageKey

Runtime storage key

error: Box<Error>

Underlying error.

FailedToReadRuntimeVersion

Failed to read runtime version of given chain.

Fields

chain: String

Name of the chain where the error has happened.

error: Box<Error>

Underlying error.

FailedToGetPendingExtrinsics

Failed to get pending extrinsics.

Fields

chain: String

Name of the chain where the error has happened.

error: Box<Error>

Underlying error.

FailedToSubmitTransaction

Failed to submit transaction.

Fields

chain: String

Name of the chain where the error has happened.

error: Box<Error>

Underlying error.

FailedStateCall

Runtime call has failed.

Fields

chain: String

Name of the chain where the error has happened.

hash: String

Hash of the block we’ve tried to call at.

method: String

Runtime API method.

arguments: Bytes

Encoded method arguments.

error: Box<Error>

Underlying error.

FailedToProveStorage

Failed to prove storage keys.

Fields

chain: String

Name of the chain where the error has happened.

hash: String

Hash of the block we’ve tried to prove keys at.

storage_keys: Vec<StorageKey>

Storage keys we have tried to prove.

error: Box<Error>

Underlying error.

FailedToSubscribeBestHeaders

Failed to subscribe to GRANDPA justifications stream.

Fields

chain: String

Name of the chain where the error has happened.

error: Box<Error>

Underlying error.

FailedToSubscribeFinalizedHeaders

Failed to subscribe to GRANDPA justifications stream.

Fields

chain: String

Name of the chain where the error has happened.

error: Box<Error>

Underlying error.

FailedToSubscribeJustifications

Failed to subscribe to GRANDPA justifications stream.

Fields

chain: String

Name of the chain where the error has happened.

error: Box<Error>

Underlying error.

UnorderedFinalizedHeaders

Headers of the chain are finalized out of order. Maybe chain has been restarted?

Fields

chain: String

Name of the chain where the error has happened.

prev_number: String

Previously finalized header number.

next_number: String

New finalized header number.

BridgePalletIsHalted

The bridge pallet is halted and all transactions will be rejected.

BridgePalletIsNotInitialized

The bridge pallet is not yet initialized and all transactions will be rejected.

TransactionInvalid(TransactionValidityError)

The Substrate transaction is invalid.

WaitingForRuntimeUpgrade

The client is configured to use newer runtime version than the connected chain uses. The client will keep waiting until chain is upgraded to given version.

Fields

chain: String

Name of the chain where the error has happened.

expected: SimpleRuntimeVersion

Expected runtime version.

actual: SimpleRuntimeVersion

Actual runtime version.

FinalityProofWeightLimitExceeded

Finality proof submission exceeds size and/or weight limits.

Fields

extras: SubmitFinalityProofCallExtras

Finality proof submission extras.

Custom(String)

Custom logic error.

Implementations

impl Error

pub fn boxed(self) -> Box<Self>

Box the error.

pub fn nested(&self) -> Option<&Self>

Returns nested error reference.

pub fn failed_to_read_header_hash_by_number<C: Chain>( number: BlockNumberOf<C>, e: Error, ) -> Self

Constructs FailedToReadHeaderHashByNumber variant.

pub fn failed_to_read_header_by_hash<C: Chain>( hash: HashOf<C>, e: Error, ) -> Self

Constructs FailedToReadHeaderByHash variant.

pub fn failed_to_read_block_by_hash<C: Chain>(hash: HashOf<C>, e: Error) -> Self

Constructs FailedToReadBlockByHash variant.

pub fn failed_to_read_best_finalized_header_hash<C: Chain>(e: Error) -> Self

Constructs FailedToReadBestFinalizedHeaderHash variant.

pub fn failed_to_read_best_header<C: Chain>(e: Error) -> Self

Constructs FailedToReadBestHeader variant.

pub fn failed_to_read_runtime_version<C: Chain>(e: Error) -> Self

Constructs FailedToReadRuntimeVersion variant.

pub fn failed_to_read_storage_value<C: Chain>( at: HashOf<C>, key: StorageKey, e: Error, ) -> Self

Constructs FailedToReadStorageValue variant.

pub fn failed_to_get_pending_extrinsics<C: Chain>(e: Error) -> Self

Constructs FailedToGetPendingExtrinsics variant.

pub fn failed_to_submit_transaction<C: Chain>(e: Error) -> Self

Constructs FailedToSubmitTransaction variant.

pub fn failed_state_call<C: Chain>( at: HashOf<C>, method: String, arguments: Bytes, e: Error, ) -> Self

Constructs FailedStateCall variant.

pub fn failed_to_prove_storage<C: Chain>( at: HashOf<C>, storage_keys: Vec<StorageKey>, e: Error, ) -> Self

Constructs FailedToProveStorage variant.

pub fn failed_to_get_system_health<C: Chain>(e: Error) -> Self

Constructs FailedToGetSystemHealth variant.

pub fn failed_to_subscribe_best_headers<C: Chain>(e: Error) -> Self

Constructs FailedToSubscribeBestHeaders variant.

pub fn failed_to_subscribe_finalized_headers<C: Chain>(e: Error) -> Self

Constructs FailedToSubscribeFinalizedHeaders variant.

pub fn failed_to_subscribe_justification<C: Chain>(e: Error) -> Self

Constructs FailedToSubscribeJustifications variant.

pub fn unordered_finalized_headers<C: Chain>( prev_number: BlockNumberOf<C>, next_number: BlockNumberOf<C>, ) -> Self

Constructs Un