Struct relay_substrate_client::Client

pub struct Client<C: Chain> { /* private fields */ }

Substrate client type.

Cloning Client is a cheap operation that only clones internal references. Different clones of the same client are guaranteed to use the same references.

Implementations

impl<C: Chain> Client<C>

pub async fn new(params: ConnectionParams) -> Self

Returns client that is able to call RPCs on Substrate node over websocket connection.

This function will keep connecting to given Substrate node until connection is established and is functional. If attempt fail, it will wait for RECONNECT_DELAY and retry again.

pub async fn try_connect(params: Arc<ConnectionParams>) -> Result<Self>

Try to connect to Substrate node over websocket. Returns Substrate RPC client if connection has been established or error otherwise.

impl<C: Chain> Client<C>

pub async fn simple_runtime_version(&self) -> Result<SimpleRuntimeVersion>

Return simple runtime version, only include spec_version and transaction_version.

pub async fn ensure_synced(&self) -> Result<()>

Returns true if client is connected to at least one peer and is in synced state.

pub fn genesis_hash(&self) -> &C::Hash

Return hash of the genesis block.

pub async fn best_finalized_header_hash(&self) -> Result<C::Hash>

Return hash of the best finalized block.

pub async fn best_finalized_header_number(&self) -> Result<C::BlockNumber>

Return number of the best finalized block.

pub async fn best_finalized_header(&self) -> Result<C::Header>

Return header of the best finalized block.

pub async fn best_header(&self) -> Result<C::Header>where C::Header: DeserializeOwned,

Returns the best Substrate header.

pub async fn get_block( &self, block_hash: Option<C::Hash> ) -> Result<C::SignedBlock>

Get a Substrate block from its hash.

pub async fn header_by_hash(&self, block_hash: C::Hash) -> Result<C::Header>where C::Header: DeserializeOwned,

Get a Substrate header by its hash.

pub async fn block_hash_by_number( &self, number: C::BlockNumber ) -> Result<C::Hash>

Get a Substrate block hash by its number.

pub async fn header_by_number( &self, block_number: C::BlockNumber ) -> Result<C::Header>where C::Header: DeserializeOwned,

Get a Substrate header by its number.

pub async fn runtime_version(&self) -> Result<RuntimeVersion>

Return runtime version.

pub async fn storage_value<T: Send + Decode + 'static>( &self, storage_key: StorageKey, block_hash: Option<C::Hash> ) -> Result<Option<T>>

Read value from runtime storage.

pub async fn storage_map_value<T: StorageMapKeyProvider>( &self, pallet_prefix: &str, key: &T::Key, block_hash: Option<C::Hash> ) -> Result<Option<T::Value>>

Read MapStorage value from runtime storage.

pub async fn storage_double_map_value<T: StorageDoubleMapKeyProvider>( &self, pallet_prefix: &str, key1: &T::Key1, key2: &T::Key2, block_hash: Option<C::Hash> ) -> Result<Option<T::Value>>

Read DoubleMapStorage value from runtime storage.

pub async fn raw_storage_value( &self, storage_key: StorageKey, block_hash: Option<C::Hash> ) -> Result<Option<StorageData>>

Read raw value from runtime storage.

pub async fn next_account_index( &self, account: C::AccountId ) -> Result<C::Nonce>

Get the nonce of the given Substrate account.

Note: It’s the caller’s responsibility to make sure account is a valid SS58 address.

pub async fn submit_unsigned_extrinsic( &self, transaction: Bytes ) -> Result<C::Hash>

Submit unsigned extrinsic for inclusion in a block.

Note: The given transaction needs to be SCALE encoded beforehand.

pub async fn submit_signed_extrinsic( &self, signer: &AccountKeyPairOf<C>, prepare_extrinsic: impl FnOnce(HeaderIdOf<C>, C::Nonce) -> Result<UnsignedTransaction<C>> + Send + 'static ) -> Result<C::Hash>where C: ChainWithTransactions, C::AccountId: From<<C::AccountKeyPair as Pair>::Public>,

Submit an extrinsic signed by given account.

All calls of this method are synchronized, so there can’t be more than one active submit_signed_extrinsic() call. This guarantees that no nonces collision may happen if all client instances are clones of the same initial Client.

Note: The given transaction needs to be SCALE encoded beforehand.

pub async fn submit_and_watch_signed_extrinsic( &self, signer: &AccountKeyPairOf<C>, prepare_extrinsic: impl FnOnce(HeaderIdOf<C>, C::Nonce) -> Result<UnsignedTransaction<C>> + Send + 'static ) -> Result<TransactionTracker<C, Self>>where C: ChainWithTransactions, C::AccountId: From<<C::AccountKeyPair as Pair>::Public>,

Does exactly the same as submit_signed_extrinsic, but keeps watching for extrinsic status after submission.

pub async fn pending_extrinsics(&self) -> Result<Vec<Bytes>>

Returns pending extrinsics from transaction pool.

pub async fn validate_transaction<SignedTransaction: Encode + Send + 'static>( &self, at_block: C::Hash, transaction: SignedTransaction ) -> Result<TransactionValidity>

Validate transaction at given block state.

pub async fn extimate_extrinsic_weight<SignedTransaction: Encode + Send + 'static>( &self, transaction: SignedTransaction ) -> Result<Weight>

Returns weight of the given transaction.

pub async fn grandpa_authorities_set( &self, block: C::Hash ) -> Result<OpaqueGrandpaAuthoritiesSet>

Get the GRANDPA authority set at given block.

pub async fn typed_state_call<Input: Encode, Output: Decode>( &self, method_name: String, input: Input, at_block: Option<C::Hash> ) -> Result<Output>

Execute runtime call at given block, provided the input and output types. It also performs the input encode and output decode.

pub async fn state_call( &self, method: String, data: Bytes, at_block: Option<C::Hash> ) -> Result<Bytes>

Execute runtime call at given block.

pub async fn prove_storage( &self, keys: Vec<StorageKey>, at_block: C::Hash ) -> Result<StorageProof>

Returns storage proof of given storage keys.

pub async fn token_decimals(&self) -> Result<Option<u64>>

Return tokenDecimals property from the set of chain properties.

pub async fn subscribe_finality_justifications<FC: SubstrateFinalityClient<C>>( &self ) -> Result<Subscription<Bytes>>

Return new finality justifications stream.

pub async fn generate_grandpa_key_ownership_proof( &self, at: HashOf<C>, set_id: SetId, authority_id: AuthorityId ) -> Result<Option<OpaqueKeyOwnershipProof>>where C: ChainWithGrandpa,

Generates a proof of key ownership for the given authority in the given set.

pub fn can_start_version_guard(&self) -> bool

Returns true if version guard can be started.

There’s no reason to run version guard when version mode is set to Auto. It can lead to relay shutdown when chain is upgraded, even though we have explicitly said that we don’t want to shutdown.