Trait relay_substrate_client::Chain

pub trait Chain: ChainBase + Clone {
    type SignedBlock: Member + Serialize + DeserializeOwned + BlockWithJustification<Self::Header>;
    type Call: Clone + Codec + Debug + Send + Sync;

    const NAME: &'static str;
    const BEST_FINALIZED_HEADER_ID_METHOD: &'static str;
    const AVERAGE_BLOCK_INTERVAL: Duration;
}

Substrate-based chain from minimal relay-client point of view.

Required Associated Types

type SignedBlock: Member + Serialize + DeserializeOwned + BlockWithJustification<Self::Header>

Block type.

type Call: Clone + Codec + Debug + Send + Sync

The aggregated Call type.

Required Associated Constants

const NAME: &'static str

Chain name.

const BEST_FINALIZED_HEADER_ID_METHOD: &'static str

Name of the runtime API method that is returning best known finalized header number and hash (as tuple).

Keep in mind that this method is normally provided by the other chain, which is bridged with this chain.

const AVERAGE_BLOCK_INTERVAL: Duration

Average block interval.

How often blocks are produced on that chain. It’s suggested to set this value to match the block time of the chain.

Object Safety

This trait is not object safe.