Trait relay_substrate_client::ChainBase

pub trait ChainBase: 'static + Send + Sync {
    type BlockNumber: Parameter + Member + MaybeSerializeDeserialize + Hash + Copy + Default + MaybeDisplay + AtLeast32BitUnsigned + FromStr + MaybeMallocSizeOf + AsPrimitive<usize>;
    type Hash: Parameter + Member + MaybeSerializeDeserialize + Hash + Ord + Copy + MaybeDisplay + Default + SimpleBitOps + AsRef<[u8]> + AsMut<[u8]> + MaybeMallocSizeOf;
    type Hasher: Hash
    where
        <Self::Hasher as Hash>::Output == Self::Hash;
    type Header: Parameter + Header + MaybeSerializeDeserialize
    where
        <Self::Header as Header>::Number == Self::BlockNumber,
        <Self::Header as Header>::Hash == Self::Hash;
}

Minimal Substrate-based chain representation that may be used from no_std environment.

Required Associated Types

type BlockNumber: Parameter + Member + MaybeSerializeDeserialize + Hash + Copy + Default + MaybeDisplay + AtLeast32BitUnsigned + FromStr + MaybeMallocSizeOf + AsPrimitive<usize>

A type that fulfills the abstract idea of what a Substrate block number is.

type Hash: Parameter + Member + MaybeSerializeDeserialize + Hash + Ord + Copy + MaybeDisplay + Default + SimpleBitOps + AsRef<[u8]> + AsMut<[u8]> + MaybeMallocSizeOf

A type that fulfills the abstract idea of what a Substrate hash is.

type Hasher: Hash
where
<Self::Hasher as Hash>::Output == Self::Hash

A type that fulfills the abstract idea of what a Substrate hasher (a type that produces hashes) is.

type Header: Parameter + Header + MaybeSerializeDeserialize
where
<Self::Header as Header>::Number == Self::BlockNumber,
<Self::Header as Header>::Hash == Self::Hash

A type that fulfills the abstract idea of what a Substrate header is.