Trait relay_substrate_client::ParachainBase

pub trait ParachainBase: Chain {
    const PARACHAIN_ID: u32;
    const MAX_HEADER_SIZE: u32;
}

Minimal parachain representation that may be used from no_std environment.

Required Associated Constants

const PARACHAIN_ID: u32

Parachain identifier.

const MAX_HEADER_SIZE: u32

Maximal size of the parachain header.

This isn’t a strict limit. The relayer may submit larger headers and the pallet will accept the call. The limit is only used to compute whether the refund can be made.

Object Safety

This trait is not object safe.