Crate relay_substrate_client

Tools to interact with Substrate node using RPC methods.

Modules

calls

Basic runtime calls.

guard

Pallet provides a set of guard functions that are running in background threads and are aborting process if some condition fails.

metrics

Contains several Substrate-specific metrics that may be exposed by relay.

test_chain

Pallet provides a set of guard functions that are running in background threads and are aborting process if some condition fails. Test chain implementation to use in tests.

Structs

ConnectionParams

Substrate-over-websocket connection params.

FullRuntimeUtilityPallet

Structure that implements UtilityPalletProvider based on a full runtime.

MockedRuntimeUtilityPallet

Structure that implements UtilityPalletProvider based on a call conversion.

SignParam

Sign transaction parameters

SimpleRuntimeVersion

A simple runtime version. It only includes the spec_version and transaction_version.

StreamDescription

Structure describing a stream.

Subscription

Subscription to some chain events.

SyncHeader

Generic wrapper for sp_runtime::traits::Header based headers, that implements finality_relay::SourceHeader and may be used in headers sync directly.

TransactionTracker

Substrate transaction tracker implementation.

UnsignedTransaction

Transaction before it is signed.

Enums

ChainRuntimeVersion

Chain runtime version in client

Error

Errors that can occur only when interacting with a Substrate node through RPC.

TransactionEra

Era of specific transaction.

Constants

ANCIENT_BLOCK_THRESHOLD

The difference between best block number and number of its ancestor, that is enough for us to consider that ancestor an “ancient” block with dropped state.

Traits

BlockWithJustification

Block with justification.

Chain

Substrate-based chain from minimal relay-client point of view.

ChainBase

Minimal Substrate-based chain representation that may be used from no_std environment.

ChainWithBalances

Substrate-based chain with AccountData generic argument of frame_system::AccountInfo set to the pallet_balances::AccountData<Balance>.

ChainWithGrandpa

Substrate-based chain that is using direct GRANDPA finality from minimal relay-client point of view.

ChainWithMessages

Substrate-based chain with messaging support from minimal relay-client point of view.

ChainWithRewards

Substrate-based chain with bridge relayers pallet as a reward ledger.

ChainWithRuntimeVersion

Bridge-supported network definition.

ChainWithTransactions

Substrate-based chain transactions signing scheme.

ChainWithUtilityPallet

Substrate-based chain that uses pallet-utility.

Client

Relay uses the Client to communicate with the node, connected to Substrate chain C.

Parachain

Substrate-based parachain from minimal relay-client point of view.

ParachainBase

Minimal parachain representation that may be used from no_std environment.

RelayChain

Substrate-based relay chain that supports parachains.

UnderlyingChainProvider

A trait that provides the type of the underlying chain.

UtilityPallet

Trait that provides functionality defined inside pallet-utility

Functions

is_ancient_block

Returns true if we think that the state is already discarded for given block.

new

Creates new RPC client with caching support.

transaction_stall_timeout

Returns stall timeout for relay loop.

Type Aliases

AccountIdOf

Account id type used by the chain.

AccountKeyPairOf

Account key pair used by transactions signing scheme.

AccountPublicOf

Account public type used by the chain.

BalanceOf

Balance type used by the chain.

BlockNumberOf

Block number used by the chain.

CallOf

Call type used by the chain.

HashOf

Hash type used by the chain.

HeaderIdOf

Header id used by the chain.

HeaderOf

Header type used by the chain.

NonceOf

Transaction nonce type used by the chain.

OpaqueGrandpaAuthoritiesSet

Opaque GRANDPA authorities set.

Result

Result type used by Substrate client.

RpcWithCachingClient

Type of RPC client with caching support.

SignatureOf

Signature type used by the chain.

SignedBlockOf

Signed block type of given chain.

TransactionEraOf

Transaction era used by the chain.

TransactionStatusOf

Transaction status of the chain.