Enum relay_substrate_client::Error

pub enum Error {

    Io(Error),
    RpcError(ClientError),
    ResponseParseFailed(Error),
    AccountDoesNotExist,
    MissingMandatoryStorageValue,
    MissingRequiredParachainHead(ParaId, u64),
    FinalityProofNotFound(u64),
    ClientNotSynced(Health),
    FailedToReadBestFinalizedHeaderHash {
        chain: String,
        error: Box<Error>,
    },
    FailedToReadBestHeader {
        chain: String,
        error: Box<Error>,
    },
    FailedToReadHeaderByHash {
        chain: String,
        hash: String,
        error: Box<Error>,
    },
    ErrorExecutingRuntimeCall {
        chain: String,
        method: String,
        error: Box<Error>,
    },
    FailedToReadRuntimeStorageValue {
        chain: String,
        key: StorageKey,
        error: Box<Error>,
    },
    BridgePalletIsHalted,
    BridgePalletIsNotInitialized,
    NoParachainHeadAtTarget(u32, String),
    StorageProofError(StorageProofError),
    TransactionInvalid(TransactionValidityError),
    WaitingForRuntimeUpgrade {
        chain: String,
        expected: SimpleRuntimeVersion,
        actual: SimpleRuntimeVersion,
    },
    Custom(String),

}

Errors that can occur only when interacting with a Substrate node through RPC.

Variants

Io(Error)

IO error.

RpcError(ClientError)

An error that can occur when making a request to an JSON-RPC server.

ResponseParseFailed(Error)

The response from the server could not be SCALE decoded.

AccountDoesNotExist

Account does not exist on the chain.

MissingMandatoryStorageValue

Runtime storage is missing some mandatory value.

MissingRequiredParachainHead(ParaId, u64)

Required parachain head is not present at the relay chain.

FinalityProofNotFound(u64)

Failed to find finality proof for the given header.

ClientNotSynced(Health)

The client weโre connected to is not synced, so we canโt rely on its state.

FailedToReadBestFinalizedHeaderHash

Fields

chain: String

Name of the chain where the error has happened.

error: Box<Error>

Underlying error.

Failed to read best finalized header hash from given chain.

FailedToReadBestHeader

Fields

chain: String

Name of the chain where the error has happened.

error: Box<Error>

Underlying error.

Failed to read best finalized header from given chain.

FailedToReadHeaderByHash

Fields

chain: String

Name of the chain where the error has happened.

hash: String

Hash of the header weโve tried to read.

error: Box<Error>

Underlying error.

Failed to read header by hash from given chain.

ErrorExecutingRuntimeCall

Fields

chain: String

Name of the chain where the error has happened.

method: String

Runtime method name.

error: Box<Error>

Underlying error.

Failed to execute runtime call at given chain.

FailedToReadRuntimeStorageValue

Fields

chain: String

Name of the chain where the error has happened.

key: StorageKey

Runtime storage key

error: Box<Error>

Underlying error.

Failed to read sotrage value at given chain.

BridgePalletIsHalted

The bridge pallet is halted and all transactions will be rejected.

BridgePalletIsNotInitialized

The bridge pallet is not yet initialized and all transactions will be rejected.

NoParachainHeadAtTarget(u32, String)

Thereโs no best head of the parachain at the pallet-bridge-parachains at the target side.

StorageProofError(StorageProofError)

An error has happened when we have tried to parse storage proof.

TransactionInvalid(TransactionValidityError)

The Substrate transaction is invalid.

WaitingForRuntimeUpgrade

Fields

chain: String

Name of the chain where the error has happened.

expected: SimpleRuntimeVersion

Expected runtime version.

actual: SimpleRuntimeVersion

Actual runtime version.

The client is configured to use newer runtime version than the connected chain uses. The client will keep waiting until chain is upgraded to given version.

Custom(String)

Custom logic error.

Implementations

impl Error

pub fn boxed(self) -> Box<Self>

Box the error.